+++ pipeline/backend/collector.go
@@ -2,7 +2,7 @@ package backend
 
 var workload = make(chan *RepoData, 100)
 
-func collector(repodata map[int]*RepoData) {
+func collector(repodata map[int64]*RepoData) {
 	if len(repodata) != 0 {
 		for _, rd := range repodata {
 			workload <- rd
